First Release
These are bash scripts for downloading OpenVAS 7 on a Ubuntu 14 server.
It may work on other systems, but I know it works on this. You should
run getopenvas.sh to download and compile all the packages. After that
completes, run firstsetupopenvas.sh to do necessary preparations.

A future version will create these as init.d services.